# dev-files/apispec_50ksZCN5.json
{"openapi":"3.0.0","info":{"title":"Blog API","description":"","version":"0.0.1"},"servers":[{"url":"https://x8ki-letl-twmt.n7.xano.io/api:50ksZCN5"}],"components":{"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}},"schemas":{}},"paths":{"/category/{category_id}":{"delete":{"summary":"Delete category record.","description":"Delete category record.\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["category"],"security":[],"parameters":[{"name":"category_id","in":"path","description":"","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}}},"get":{"summary":"Get category record","description":"Get category record\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["category"],"security":[],"parameters":[{"name":"category_id","in":"path","description":"","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}}},"patch":{"summary":"Edit category record","description":"Edit category record\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["category"],"security":[],"parameters":[{"name":"category_id","in":"path","description":"","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}},"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}}}}}},"/category":{"get":{"summary":"Query all category records","description":"Query all category records\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["category"],"security":[],"parameters":[],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}}},"post":{"summary":"Add category record","description":"Add category record\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["category"],"security":[],"parameters":[],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}},"multipart/form-data":{"schema":{"type":"object","properties":{"name":{"type":"string","description":""},"description":{"type":"string","description":""},"color":{"type":"string","description":""}}}}}}}},"/post/{post_id}":{"delete":{"summary":"Delete post record.","description":"Delete post record.\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["post"],"security":[],"parameters":[{"name":"post_id","in":"path","description":"","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}}},"get":{"summary":"Get post record","description":"Get post record\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["post"],"security":[],"parameters":[{"name":"post_id","in":"path","description":"","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}}},"patch":{"summary":"Edit post record","description":"Edit post record\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["post"],"security":[],"parameters":[{"name":"post_id","in":"path","description":"","required":true,"schema":{"type":"string","format":"uuid"}}],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}},"multipart/form-data":{"schema":{"type":"object","properties":{"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}}}}}},"/post":{"get":{"summary":"Query all post records","description":"Query all post records\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["post"],"security":[],"parameters":[],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"array","items":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}}},"post":{"summary":"Add post record","description":"Add post record\n\u003Cbr /\u003E\u003Cbr /\u003E\n\u003Cb\u003EAuthentication:\u003C/b\u003E not required","tags":["post"],"security":[],"parameters":[],"responses":{"200":{"description":"Success!","content":{"application/json":{"schema":{"type":"object","properties":{"id":{"type":"string","format":"uuid","description":""},"created_at":{"type":"number","format":"timestamptz","description":"","default":"now"},"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}}}},"400":{"description":"Input Error. Check the request payload for issues."},"401":{"description":"Unauthorized"},"403":{"description":"Access denied. Additional privileges are needed access the requested resource."},"404":{"description":"Not Found. The requested resource does not exist."},"429":{"description":"Rate Limited. Too many requests."},"500":{"description":"Unexpected error"}},"requestBody":{"content":{"application/json":{"schema":{"type":"object","properties":{"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}},"multipart/form-data":{"schema":{"type":"object","properties":{"title":{"type":"string","description":""},"content":{"type":"string","description":""},"published":{"type":"boolean","description":"","default":"false"},"user_id":{"type":"string","format":"uuid","description":""}}}}}}}}}}