# src/app/api/category/route.ts
import { NextResponse } from 'next/server'
import { crudService } from '@/lib/api/crud-service'

export async function GET() {
  try {
    const categories = await crudService.getAll('category')
    return NextResponse.json(categories)
  } catch (error) {
    console.error('Failed to fetch categories:', error)
    return NextResponse.json(
      { error: 'Failed to fetch categories' },
      { status: 500 }
    )
  }
}

export async function POST(request: Request) {
  try {
    const data = await request.json()
    const category = await crudService.create('category', data)
    return NextResponse.json(category)
  } catch (error) {
    console.error('Failed to create category:', error)
    return NextResponse.json(
      { error: 'Failed to create category' },
      { status: 500 }
    )
  }
}