# src/app/api/post/route.ts
import { NextResponse } from 'next/server'
import { crudService } from '@/lib/api/crud-service'

export async function GET() {
  try {
    const posts = await crudService.getAll('post')
    return NextResponse.json(posts)
  } catch (error) {
    console.error('Failed to fetch posts:', error)
    return NextResponse.json(
      { error: 'Failed to fetch posts' },
      { status: 500 }
    )
  }
}

export async function POST(request: Request) {
  try {
    const data = await request.json()
    const post = await crudService.create('post', data)
    return NextResponse.json(post)
  } catch (error) {
    console.error('Failed to create post:', error)
    return NextResponse.json(
      { error: 'Failed to create post' },
      { status: 500 }
    )
  }
}